requirements.txt
fastapi
uvicorn
python-multipart
huggingface_hub
chromadb
pypdf
edge-tts
aiofiles
jinja2
sentence-transformers
python-docx